feat: support base directory

+++ encfs/fs.go
@@ -2,6 +2,8 @@ package encfs
 
 import (
 	"os"
+	"path"
+	"path/filepath"
 	"strings"
 	"time"
 
@@ -21,18 +23,27 @@ func NewEncryptionMasterKey(key []byte) *EncryptionMasterKey {
 }
 
 type EncFs struct {
-	key *EncryptionMasterKey
+	baseDirectory string
+	key           *EncryptionMasterKey
 }
 
 func NewEncFs(key *EncryptionMasterKey) afero.Fs {
+	return NewEncFsWithBaseDirecotry("", key)
+}
+
+func NewEncFsWithBaseDirecotry(baseDirectory string, key *EncryptionMasterKey) afero.Fs {
 	return &EncFs{
-		key: key,
+		baseDirectory,
+		key,
 	}
 }
 
 func (*EncFs) Name() string { return "EncFs" }
 
 func (encFs *EncFs) Create(name string) (afero.File, error) {
+	if name = encFs.resolve(name); name == "" {
+		return nil, os.ErrNotExist
+	}
 	if err := checkFileExt(name); err != nil {
 		return nil, err
 	}
@@ -45,15 +56,28 @@ func (encFs *EncFs) Create(name string) (afero.File, error) {
 	return convertOsFileToEncFile(name, f, e, encFs, true)
 }
 
-func (*EncFs) Mkdir(name string, perm os.FileMode) error {
+func (encFs *EncFs) Mkdir(name string, perm os.FileMode) error {
+	if name = encFs.resolve(name); name == "" {
+		return os.ErrNotExist
+	}
 	return os.Mkdir(name, perm)
 }
 
-func (*EncFs) MkdirAll(path string, perm os.FileMode) error {
+func (encFs *EncFs) MkdirAll(path string, perm os.FileMode) error {
+	if path = encFs.resolve(path); path == "" {
+		return os.ErrNotExist
+	}
+	if path == filepath.Clean(encFs.baseDirectory) {
+		// Prohibit removing the virtual root directory.
+		return os.ErrInvalid
+	}
 	return os.MkdirAll(path, perm)
 }
 
 func (encFs *EncFs) Open(name string) (afero.File, error) {
+	if name = encFs.resolve(name); name == "" {
+		return nil, os.ErrNotExist
+	}
 	if err := checkFileExt(name); err != nil {
 		return nil, err
 	}
@@ -67,6 +91,9 @@ func (encFs *EncFs) Open(name string) (afero.File, error) {
 }
 
 func (encFs *EncFs) OpenFile(name string, flag int, perm os.FileMode) (afero.File, error) {
+	if name = encFs.resolve(name); name == "" {
+		return nil, os.ErrNotExist
+	}
 	if err := checkFileExt(name); err != nil {
 		return nil, err
 	}
@@ -79,13 +106,23 @@ func (encFs *EncFs) OpenFile(name string, flag int, perm os.FileMode) (afero.Fil
 	return convertOsFileToEncFile(name, f, e, encFs, false)
 }
 
-func (*EncFs) Remove(name string) error {
+func (encFs *EncFs) Remove(name string) error {
+	if name = encFs.resolve(name); name == "" {
+		return os.ErrNotExist
+	}
 	encFileMetaName := name + EncFileExt
 	_ = os.Remove(encFileMetaName)
 	return os.Remove(name)
 }
 
 func (encFs *EncFs) RemoveAll(path string) error {
+	if path = encFs.resolve(path); path == "" {
+		return os.ErrNotExist
+	}
+	if path == filepath.Clean(encFs.baseDirectory) {
+		// Prohibit removing the virtual root directory.
+		return os.ErrInvalid
+	}
 	fileInfo, err := os.Stat(path)
 	if err == nil && !fileInfo.IsDir() {
 		return encFs.Remove(path)
@@ -93,40 +130,95 @@ func (encFs *EncFs) RemoveAll(path string) error {
 	return os.RemoveAll(path)
 }
 
-func (*EncFs) Rename(oldname, newname string) error {
+func (encFs *EncFs) Rename(oldname, newname string) error {
+	if oldname = encFs.resolve(oldname); oldname == "" {
+		return os.ErrNotExist
+	}
+	if newname = encFs.resolve(newname); newname == "" {
+		return os.ErrNotExist
+	}
+	if root := filepath.Clean(encFs.baseDirectory); root == oldname || root == newname {
+		// Prohibit renaming from or to the virtual root directory.
+		return os.ErrInvalid
+	}
 	oldEncFileMetaName := oldname + EncFileExt
 	newEncFileMetaName := newname + EncFileExt
 	_ = os.Rename(oldEncFileMetaName, newEncFileMetaName)
 	return os.Rename(oldname, newname)
 }
 
-func (*EncFs) Stat(name string) (os.FileInfo, error) {
+func (encFs *EncFs) Stat(name string) (os.FileInfo, error) {
+	if name = encFs.resolve(name); name == "" {
+		return nil, os.ErrNotExist
+	}
 	return os.Stat(name)
 }
 
-func (*EncFs) Chmod(name string, mode os.FileMode) error {
+func (encFs *EncFs) Chmod(name string, mode os.FileMode) error {
+	if name = encFs.resolve(name); name == "" {
+		return os.ErrNotExist
+	}
 	return os.Chmod(name, mode)
 }
 
-func (*EncFs) Chown(name string, uid, gid int) error {
+func (encFs *EncFs) Chown(name string, uid, gid int) error {
+	if name = encFs.resolve(name); name == "" {
+		return os.ErrNotExist
+	}
 	return os.Chown(name, uid, gid)
 }
 
-func (*EncFs) Chtimes(name string, atime time.Time, mtime time.Time) error {
+func (encFs *EncFs) Chtimes(name string, atime time.Time, mtime time.Time) error {
+	if name = encFs.resolve(name); name == "" {
+		return os.ErrNotExist
+	}
 	return os.Chtimes(name, atime, mtime)
 }
 
-func (*EncFs) LstatIfPossible(name string) (os.FileInfo, bool, error) {
+func (encFs *EncFs) LstatIfPossible(name string) (os.FileInfo, bool, error) {
+	if name = encFs.resolve(name); name == "" {
+		return nil, false, os.ErrNotExist
+	}
 	fi, err := os.Lstat(name)
 	return fi, true, err
 }
 
-func (*EncFs) SymlinkIfPossible(oldname, newname string) error {
+func (encFs *EncFs) SymlinkIfPossible(oldname, newname string) error {
+	if oldname = encFs.resolve(oldname); oldname == "" {
+		return os.ErrNotExist
+	}
+	if newname = encFs.resolve(newname); newname == "" {
+		return os.ErrNotExist
+	}
+	if root := filepath.Clean(encFs.baseDirectory); root == oldname || root == newname {
+		// Prohibit symlink from or to the virtual root directory.
+		return os.ErrInvalid
+	}
 	return os.Symlink(oldname, newname)
 }
 
-func (*EncFs) ReadlinkIfPossible(name string) (string, error) {
+func (encFs *EncFs) ReadlinkIfPossible(name string) (string, error) {
+	if name = encFs.resolve(name); name == "" {
+		return "", os.ErrNotExist
+	}
 	return os.Readlink(name)
+}
+
+// copy from: golang.org/x/net/webdav/file.go
+func (encFs *EncFs) resolve(name string) string {
+	if encFs.baseDirectory == "" {
+		return name
+	}
+	// This implementation is based on Dir.Open's code in the standard net/http package.
+	if filepath.Separator != '/' && strings.IndexRune(name, filepath.Separator) >= 0 ||
+		strings.Contains(name, "\x00") {
+		return ""
+	}
+	dir := string(encFs.baseDirectory)
+	if dir == "" {
+		dir = "."
+	}
+	return filepath.Join(dir, filepath.FromSlash(slashClean(name)))
 }
 
 func checkFileExt(name string) error {
@@ -145,4 +237,14 @@ func convertOsFileToEncFile(name string, file *os.File, e error, encFs *EncFs, i
 		return nil, err
 	}
 	return encFile, nil
+}
+
+// copy from: golang.org/x/net/webdav/file.go
+// slashClean is equivalent to but slightly more efficient than
+// path.Clean("/" + name).
+func slashClean(name string) string {
+	if name == "" || name[0] != '/' {
+		name = "/" + name
+	}
+	return path.Clean(name)
 }
